[wx] Disable dragging status bar buttons
Mouse capture doesn't quite work the same as it used to.
It works fine for the notebook tabs, but for the buttons
on the status bar it doesn't.  It appears the button OnClick
event is firing, causing a forced mouse capture lost.  On
older wxPython, capturing the mouse actually disabled any
click events from being generated.