Rework repo cleaning for building
Removed all those file shenanigans in favour of simple git stashing for
ensuring a clean repo when building the distributables.

We should get a slight performance boost since we're no longer iterating
over the entire repository 50 times per build - the stashing is slow,
yes, but it makes up for it by making everything else lightning fast.

File lists are dumb and no longer used for 7z packaging - 7z has a very
good wildcard/exclusion system and we should be using that.

Printing the python version and those git warnings was kept, it's not
super useful but pretty adorable.

Removed the ability to change logfiles and output folder - no one was
using the anyway and makes the stashing much more straightforward.

The loot api is no longer imported directly but only as-needed -
importing it at the start blocks stashing since the files are being
used. A simple check if the files exist is enough to check whether
they're installed.