NotebookPanel(WithFirstShow, PanelWin) XXXXXXX
TTT check basher.INIPanel.ShowPanel - shows okay
despite RefreshUI runs after super.ShowPanel - because of calling
PopulateItems() at the very end balt.UIList.__init__ - note that some
panels like installers are not refreshed at this moment so this doesn't
do anything - why does not work for mods/saves, those are initialized
Still looking for a way to use EVT_SHOW

Turns out EVT_SHOW is triggered on AddPage for
Installers (first tab to add) and on self.SetSelection(pageIndex) for
mods tab (the one that shows first in my case here). The scrollbar is set
correctly (as modInfos is refreshed) but detailsPanel is not set the
sashPos. Plus tabs (like installers) that are not refreshed will not
correctly update scollbars etc ever - so reverting (plus probably we
must check if it's a show or hide event I guess, those are triggered too)
Would be nice to have this really on show though - note simplification
of ShowPanel