structs_cache WIP:
This is clearly not for performance although it might help in cases
like this:

@@ -971,3 +963,3 @@ def load_frag(self, record, ins, vmad_version, obj_format, read_id,
                 record.prop_data = [x for x in ins.unpack(
-                    struct.Struct(u'%df' % array_len).unpack, array_len * 4,
+                    structs_cache[u'%df' % array_len].unpack, array_len * 4,
                     read_id)]

where we created those on the fly. It is a preparation for what is to
come - we need to centralize Srtuct.

Mopy/bash/bosh/_saves.py:
Mopy/bash/bosh/bsa_files.py:
Mopy/bash/bosh/save_headers.py:
struct_pack pruning

Note I left imports unformatted to avoid conflicts