Upgrade to wxPython 4.1 RRR
Upgrading for a few reasons:
 - Allows us to drop the PNG optimization step before running a wizard.
   Note that I kept pngcrush around - it's stuck in our history either
   way, so I'd rather put it to use by adding an option for optimizing
   PNGs to BAIN projects.
 - Keeping up with breaking wx changes is a good idea (e.g. the
   alignment flags change).
 - wx.ListCtrl (which is probably the most important widget we use) got
   a major rework in 4.1.x by becoming native. It looks quite different
   and we should get testing done to see if anything breaks with it.

Note that 4.1.0 is the last py2 version of wxPython, 4.1.1 is already
py3 only.

wx4.1 build fixes

Whew, adding yet more DLLs to this stupid field...

Add PyMuPDF to requirements.txt

Still completely optional, but recommended since it allows the Doc
Browser to display PDFs as well. The wxPython pdfViewer was broken on
wx4.0, but works fine on wx4.1.

Note: newer versions of PyMuPDF advertise that they are compatible with
py2 on pypi, but 1.18.0 is the last version for which wheels are
available - and manually building this lib is an absolute pain.

Closes # 553 <--- RRR