Initial high DPI awareness RRR
Basically just means marking ourselves as aware on Windows and only
assigning size/position after creating the widget so that we can treat
all our pixel specifications as device-independent pixels (DIP).

However, that caused the game select popup to become misaligned and ugly
- because it uses raw wx, of course. This turned into a wild goose chase
and ended up with me rewriting all wx usages in bash.py and de-wx'ing a
lot of it.

Remaining work is creating high-resolution versions of all the icons we
use in our GUI, of course.

Under # 190 for de-wx'ing of bash.py
Closes # 555, for now - over the course of # 190 we'll want to <--- RRR
create SVG versions of all our icons so we can get crisp,
high-resolution images in our GUI without having to include tons of
different sizes.