Adjust window size when it's too small
SetMinSize only applies to when the user resizes the window. It does
nothing to the initial size we pass in. So let's do that manually
instead.

Now when you adjust the _min_size of a window, the next time the window
is opened, if the width or height is too small, that dimension will be
bumped up to match the one specified in _min_size.