Drop special ComboBox unsubscribe handling
This is fixed in wx4.1 - no more tracebacks when we don't specifically
unsubscribe from ComboBox's EVT_SIZE.

Mopy/bash/balt.py:
Correct temporary subscribe/unsubscribe usage

The right pattern for this is the same as that for locks:

    begin_section()
    try:
        other_code()
    finally:
        end_section()

That way, if the begin_section call fails, we exit the entire method.
The previous code did it like this:

    try:
        begin_section()
	other_code()
    finally:
        end_section()

Here, if begin_section fails, we still enter the finally. Which will now
fail as well because we never began the section to begin with, leading
to an incorrect error message.

In the case of locks, begin_section is Lock.acquire and end_section is
Lock.release. In the case of GUI components, begin_section is
EventHandler.unsubscribe and end_section is EventHandler.subscribe.